Backed out 3 changesets (bug 1724368, bug 1725154, bug 1724388) for causing bustages in minidump_processor.cc
CLOSED TREE

Backed out changeset 8be159457667 (bug 1724388)
Backed out changeset fd5d00822477 (bug 1724368)
Backed out changeset 092785e2a7f8 (bug 1725154)

UltraBlame original commit: 35aecc2db94d6be53e02aa88f1d54575b1782bc5

+++ toolkit/crashreporter/breakpad-client/mac/handler/exception_handler.h
@@ -165,8 +165,8 @@ class ExceptionHandler {
 
   
   static bool WriteForwardedExceptionMinidump(int exception_type,
-					      int64_t exception_code,
-					      int64_t exception_subcode,
+					      int exception_code,
+					      int exception_subcode,
 					      mach_port_t thread,
 					      mach_port_t task);
 
@@ -202,7 +202,7 @@ class ExceptionHandler {
   
   
   bool WriteMinidumpWithException(int exception_type,
-                                  int64_t exception_code,
+                                  int exception_code,
                                   int64_t exception_subcode,
                                   breakpad_ucontext_t *task_context,
                                   mach_port_t thread_name,
